WIP starting on godot implementation

+++ src/Game/DeploymentOrder.cs
@@ -5,5 +5,5 @@ public class DeploymentOrder
     public ulong TargetId;
     public int WarriorCount;
     public int ArcherCount;
-    public float WaveCooldown;
+    public double WaveCooldown;
 }
